sole trader refactoring
