+++ src/main/controllers/index.ts
@@ -1,5 +1,8 @@
+export * as indexController from "./indexController";
+export * as healthCheckController from "./healthCheckController";
+export * as accessibilityStatementController from "./accessibilityStatementController";
+
 // Sole trader
-export * as indexController from "./features/SoleTrader/indexController";
 export * as soleTraderDateOfBirthController from "./features/SoleTrader/dateOfBirthController";
 export * as soleTraderNameController from "./features/SoleTrader/nameController";
 export * as soleTraderWhatIsYourRoleController from "./features/SoleTrader/whatIsYourRoleController";
@@ -7,26 +10,27 @@ export * as soleTraderWhereDoYouLiveController from "./features/SoleTrader/where
 export * as soleTraderCorrespondenceAddressManualController from "./features/SoleTrader/correspondenceAddressManualController";
 export * as soleTraderCorrespondenceAddressAutoLookupController from "./features/SoleTrader/correspondenceAddressAutoLookupController";
 export * as soleTraderSectorYouWorkInController from "./features/SoleTrader/sectorYouWorkInController";
-export * as whichSectorOtherController from "./features/SoleTrader/whichSectorOtherController";
+export * as soleTraderWhichSectorOtherController from "./features/SoleTrader/WhichSectorOtherController";
 export * as soleTraderCorrespodanceAddressDetailsController from "./features/SoleTrader/correspodanceAddressDetailsController";
 export * as soleTraderCorrespondenceAddressConfirmController from "./features/SoleTrader/correspondenceAddressConfirmController";
 export * as soleTraderNationalityController from "./features/SoleTrader/nationalityController";
-export * as typeOfBusinessController from "./features/SoleTrader/typeOfBusinessController";
-export * as OtherTypeOfBusinessController from "./features/SoleTrader/otherTypeOfBusinessController";
+export * as soleTraderTypeOfBusinessController from "./features/SoleTrader/typeOfBusinessController";
+export * as soleTraderOtherTypeOfBusinessController from "./features/SoleTrader/otherTypeOfBusinessController";
 
-export * as healthCheckController from "./healthCheckController";
+// limited
 export * as companyLookupController from "./features/limited/companyLookupController";
 export * as nameRegisteredWithAmlController from "./features/limited/nameRegisteredWithAmlController";
 export * as businessMustbeAmlRegisteredController from "./features/limited/businessMustbeAmlRegisteredController";
 export * as companyInactiveController from "./features/limited/companyInactiveController";
+export * as limitedWhatIsYourRoleController from "./features/limited/limitedWhatIsYourRoleController";
+export * as limitedSectorYouWorkInController from "./features/limited/limitedSectorYouWorkInController";
+export * as isThisYourCompanyController from "./features/limited/isThisYourCompanyController";
+
+// unincorporated
 export * as whatIsTheBusinessNameController from "./features/unincorporated/whatIsTheBusinessNameController";
-export * as accessibilityStatementController from "./accessibilityStatementController";
 export * as soleTraderWhatIsTheBusinessNameController from "./features/SoleTrader/whatIsTheBusinessNameController";
 export * as unincorporatedWhatIsYourNameController from "./features/unincorporated/unincorporatedWhatIsYourNameController";
 export * as unincorporatedNameRegisteredWithAmlController from "./features/unincorporated/unincorporatedNameRegisteredWithAmlController";
 export * as unincorporatedWhatIsYourRoleController from "./features/unincorporated/unincorporatedWhatIsYourRoleController";
 export * as stopNotRelevantOfficerController from "./features/common/stopNotRelevantOfficerController";
-export * as limitedWhatIsYourRoleController from "./features/limited/limitedWhatIsYourRoleController";
-export * as limitedSectorYouWorkInController from "./features/limited/limitedSectorYouWorkInController";
 export * as unincorporatedSectorYouWorkInController from "./features/unincorporated/unincorporatedSectorYouWorkInController";
-export * as isThisYourCompanyController from "./features/limited/isThisYourCompanyController";
