+++ src/main/controllers/indexController.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from "express";
-import * as config from "../../../config";
-import { BASE_URL, TYPE_OF_BUSINESS } from "../../../types/pageURL";
-import { addLangToUrl, getLocaleInfo, getLocalesService, selectLang } from "../../../utils/localise";
+import * as config from "../config";
+import { BASE_URL, TYPE_OF_BUSINESS } from "../types/pageURL";
+import { addLangToUrl, getLocaleInfo, getLocalesService, selectLang } from "../utils/localise";
 
 export const get = async (req: Request, res: Response, next: NextFunction) => {
     const lang = selectLang(req.query.lang);
